Move balancer vault contract into efi/balancer.ts since it references our typechain and addressesjson
