+++ src/efi-ui/balancer/useSwap/useSwap.ts
@@ -2,7 +2,7 @@ import { Vault } from "@elementfi/core-typechain";
 import { FundManagement } from "integrations/efi-balancer/FundManagement";
 import { SingleSwap } from "integrations/efi-balancer/SingleSwap";
 import { SwapKind } from "integrations/efi-balancer/SwapKind";
-import { balancerVaultContract } from "integrations/efi-balancer/vault";
+import { balancerVaultContract } from "efi/balancer/vault";
 import {
   AppToaster,
   makeErrorToast,
